Moving some http auth stuff up into api package
