+++ src/app/beer_garden/api/http/base_handler.py
@@ -23,7 +23,8 @@ from tornado.web import HTTPError, RequestHandler
 
 import beer_garden.api.http
 import beer_garden.db.mongo.models
-from beer_garden.api.http.authorization import AuthMixin, coalesce_permissions
+from beer_garden.api.http.authorization import AuthMixin
+from beer_garden.api.auth import coalesce_permissions
 from beer_garden.api.http.client import ExecutorClient
 from beer_garden.api.http.metrics import http_api_latency_total
