+++ src/app/beer_garden/api/http/handlers/v1/admin.py
@@ -2,7 +2,7 @@
 from brewtils.errors import ModelValidationError
 from brewtils.schema_parser import SchemaParser
 
-from beer_garden.api.http.authorization import check_permission, Permissions
+from beer_garden.api.auth import Permissions, check_permission
 from beer_garden.api.http.base_handler import BaseHandler
