+++ src/app/beer_garden/api/http/handlers/v1/command.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from beer_garden.api.http.authorization import authenticated, Permissions
+from beer_garden.api.auth import Permissions, authenticated
 from beer_garden.api.http.base_handler import BaseHandler
