+++ src/app/beer_garden/api/http/handlers/v1/instance.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from brewtils.schema_parser import SchemaParser
 
-from beer_garden.api.http.authorization import authenticated, Permissions
+from beer_garden.api.auth import Permissions, authenticated
 from beer_garden.api.http.base_handler import BaseHandler
