+++ src/app/beer_garden/api/http/handlers/v1/request.py
@@ -7,7 +7,7 @@ from brewtils.models import Request
 from brewtils.schema_parser import SchemaParser
 
 import beer_garden.db.api as db
-from beer_garden.api.http.authorization import authenticated, Permissions
+from beer_garden.api.auth import Permissions, authenticated
 from beer_garden.api.http.base_handler import BaseHandler
