+++ src/app/beer_garden/api/http/handlers/v1/role.py
@@ -6,9 +6,8 @@ from beer_garden.db.mongo.models import Role
 from beer_garden.db.mongo.parser import MongoParser
 from beer_garden.api.http.authorization import (
     anonymous_principal,
-    authenticated,
-    Permissions,
 )
+from beer_garden.api.auth import Permissions, authenticated
 from beer_garden.api.http.base_handler import BaseHandler
 from brewtils.errors import ModelValidationError
