+++ src/app/beer_garden/api/http/handlers/v1/system.py
@@ -2,7 +2,7 @@
 from brewtils.schema_parser import SchemaParser
 from brewtils.schemas import SystemSchema
 
-from beer_garden.api.http.authorization import authenticated, Permissions
+from beer_garden.api.auth import Permissions, authenticated
 from beer_garden.api.http.base_handler import BaseHandler
