+++ src/app/beer_garden/api/http/handlers/v1/user.py
@@ -4,12 +4,7 @@ import json
 from brewtils.schema_parser import SchemaParser
 
 import beer_garden.api.http
-from beer_garden.api.http.authorization import (
-    authenticated,
-    check_permission,
-    Permissions,
-    coalesce_permissions,
-)
+from beer_garden.api.auth import Permissions, authenticated, check_permission, coalesce_permissions
 from beer_garden.api.http.base_handler import BaseHandler
