+++ src/app/beer_garden/api/http/handlers/v2/instance.py
@@ -1,4 +1,4 @@
-from beer_garden.api.http.authorization import authenticated, Permissions
+from beer_garden.api.auth import Permissions, authenticated
 from beer_garden.api.http.base_handler import BaseHandler
 from beer_garden.api.http.client import ExecutorClient
