+++ src/app/beer_garden/api/http/handlers/v2/system.py
@@ -3,7 +3,7 @@ from brewtils.errors import ConflictError
 from brewtils.schemas import SystemSchema
 
 import beer_garden.bg_utils
-from beer_garden.api.http.authorization import authenticated, Permissions
+from beer_garden.api.auth import Permissions, authenticated
 from beer_garden.api.http.base_handler import BaseHandler
 from beer_garden.api.http.client import ExecutorClient
